+++ lib/fun_with_flags/application.ex
@@ -30,9 +30,10 @@ defmodule FunWithFlags.Application do
       Config.change_notifications_enabled? && Config.notifications_adapter.worker_spec
     rescue
       e in [UndefinedFunctionError] ->
-        Logger.error "FunWithFlags: Looks like you're trying to use #{Config.notifications_adapter} for notifications, but you haven't added its optional dependency to the Mixfile."
-        Logger.error "FunWithFlags: Optionally notifications can be disabled to exclude this dependency."
-        raise e
+        Logger.error "FunWithFlags: Looks like you're trying to use #{Config.notifications_adapter} " <>
+         "for notifications, but you haven't added its optional dependency to the Mixfile. Optionally " <>
+         "notifications can be disabled to exclude this dependency."
+        reraise e, System.stacktrace
     end
   end